+++ zenoh/src/api/builders/mod.rs
@@ -16,3 +16,4 @@ pub(crate) mod publisher;
 pub(crate) mod queryable;
 pub(crate) mod reply;
 pub(crate) mod sample;
+pub(crate) mod subscriber;
